Refactor - move classes to packages and change access to some classes
