+++ src/main/java/pl/com/bohdziewicz/magicNumberApp/MainApp.java
@@ -1,5 +1,8 @@
 package pl.com.bohdziewicz.magicNumberApp;
 
+import pl.com.bohdziewicz.magicNumberApp.testers.FileTester;
+import pl.com.bohdziewicz.magicNumberApp.testers.WhatKindOfFileFinder;
+
 class MainApp {
 
     void printResultOfTestingFile(FileToServe fileToServe) {
@@ -14,7 +17,6 @@ class MainApp {
             System.out.println("My algorithm says that the file has not correct extensions");
             System.out.println("But wait...");
             whatKindOfFileFinder.infoAboutFile(fileToServe);
-
         }
     }
 }
