+++ src/main/java/pl/com/bohdziewicz/magicNumberApp/testers/FileTesterInterface.java
@@ -1,4 +1,6 @@
-package pl.com.bohdziewicz.magicNumberApp;
+package pl.com.bohdziewicz.magicNumberApp.testers;
+
+import pl.com.bohdziewicz.magicNumberApp.FileToServe;
 
 public interface FileTesterInterface {
